add BookService implementation + fix rounding issue after first test

+++ src/main/java/be/bnp/kata/api/resources/book/response/BookBasketResponse.java
@@ -9,6 +9,6 @@ import lombok.Setter;
 @Setter
 public final class BookBasketResponse {
 
-    private int total;
+    private double total;
     
 }
